BT/requirements.txt
asgiref
certifi
charset-normalizer
idna
urllib3
crispy-bootstrap5==0.7
Django==4.2.1
django-crispy-forms==2.0
django-model-utils==4.3.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
mysql-connector-python==8.0.33
Pillow==9.5.0
PyMySQL==1.0.3
requests==2.30.0
sqlparse==0.4.4
transbank-sdk==4.0.0
xhtml2pdf